+++ src/app/(detail)/leports/page.tsx
@@ -1,0 +1,7 @@
+import { TourismList } from "@/components/maindetail/TourismList";
+
+const Leports = () => {
+  return <TourismList contentTypeId={28} title="레포츠 추천" />;
+};
+
+export default Leports;
