+++ src/app/(detail)/travel/page.tsx
@@ -1,79 +1,7 @@
-"use client";
-
-import { FetchTravel } from "@/app/api/main/Tour/AllFetch/travel/route";
-import { Loading } from "@/components/common/Loading";
-
-import { DetailTitle } from "@/components/maindetail/DetailTitle";
-import { ScrollToTopButton } from "@/components/maindetail/ScrollToTopButton";
-import { SearchBar } from "@/components/maindetail/SearchBar";
-import { TravelCard } from "@/components/maindetail/TravelCard";
-import { ApiInformation } from "@/types/Main";
-import { useQuery } from "@tanstack/react-query";
-import { useState, useMemo, useEffect } from "react";
-import { useInView } from "react-intersection-observer";
+import { TourismList } from "@/components/maindetail/TourismList";
 
 const Travel = () => {
-  const [displayCount, setDisplayCount] = useState(10);
-  const [searchTerm, setSearchTerm] = useState("");
-  const { ref, inView } = useInView();
-
-  const {
-    data: travel,
-    isPending,
-    error,
-  } = useQuery<ApiInformation[], Error>({
-    queryKey: ["travel"],
-    queryFn: FetchTravel,
-  });
-
-  const filteredTravel = useMemo(() => {
-    if (!travel) return [];
-    return travel
-      .filter((item) => item.title.toLowerCase().includes(searchTerm.toLowerCase()))
-      .sort((a, b) => {
-        if (a.firstimage && !b.firstimage) return -1;
-        if (!a.firstimage && b.firstimage) return 1;
-        return 0;
-      });
-  }, [travel, searchTerm]);
-
-  const displayedTravel = useMemo(() => {
-    return filteredTravel.slice(0, displayCount);
-  }, [filteredTravel, displayCount]);
-
-  useEffect(() => {
-    if (inView) {
-      setDisplayCount((prevCount) => prevCount + 10);
-    }
-  }, [inView]);
-
-  if (isPending) return <Loading />;
-  if (error) return <div>Error: {error.message}</div>;
-
-  return (
-    <>
-      <SearchBar searchTerm={searchTerm} setSearchTerm={setSearchTerm} />
-
-      <div className="container mx-auto px-4 py-8 relative">
-        <div className="flex my-6 gap-3">
-          <DetailTitle />
-
-          <h3>여행지 추천</h3>
-        </div>
-        <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
-          {displayedTravel.map((item) => (
-            <TravelCard key={item.contentid} item={item} />
-          ))}
-        </div>
-        {displayedTravel.length < filteredTravel.length && (
-          <div ref={ref} className="py-4 text-center">
-            Loading more...
-          </div>
-        )}
-        <ScrollToTopButton />
-      </div>
-    </>
-  );
+  return <TourismList contentTypeId={12} title="관광지 추천" />;
 };
 
 export default Travel;
