--- src/app/(detail)/travelCourse/page.tsx
@@ -1,79 +1,0 @@
-"use client";
-
-import { FetchTravelCourse } from "@/app/api/main/Tour/AllFetch/travelcourse/route";
-import { Loading } from "@/components/common/Loading";
-
-import { DetailTitle } from "@/components/maindetail/DetailTitle";
-import { ScrollToTopButton } from "@/components/maindetail/ScrollToTopButton";
-import { SearchBar } from "@/components/maindetail/SearchBar";
-import { TravelCard } from "@/components/maindetail/TravelCard";
-import { ApiInformation } from "@/types/Main";
-import { useQuery } from "@tanstack/react-query";
-import { useState, useMemo, useEffect } from "react";
-import { useInView } from "react-intersection-observer";
-
-const TravelCourse = () => {
-  const [displayCount, setDisplayCount] = useState(10);
-  const [searchTerm, setSearchTerm] = useState("");
-  const { ref, inView } = useInView();
-
-  const {
-    data: travelCourse,
-    isPending,
-    error,
-  } = useQuery<ApiInformation[], Error>({
-    queryKey: ["travelCourse"],
-    queryFn: FetchTravelCourse,
-  });
-
-  const filteredTravelCourse = useMemo(() => {
-    if (!travelCourse) return [];
-    return travelCourse
-      .filter((item) => item.title.toLowerCase().includes(searchTerm.toLowerCase()))
-      .sort((a, b) => {
-        if (a.firstimage && !b.firstimage) return -1;
-        if (!a.firstimage && b.firstimage) return 1;
-        return 0;
-      });
-  }, [travelCourse, searchTerm]);
-
-  const displayedTravelCourse = useMemo(() => {
-    return filteredTravelCourse.slice(0, displayCount);
-  }, [filteredTravelCourse, displayCount]);
-
-  useEffect(() => {
-    if (inView) {
-      setDisplayCount((prevCount) => prevCount + 10);
-    }
-  }, [inView]);
-
-  if (isPending) return <Loading />;
-  if (error) return <div>Error: {error.message}</div>;
-
-  return (
-    <>
-      <SearchBar searchTerm={searchTerm} setSearchTerm={setSearchTerm} />
-
-      <div className="container mx-auto px-4 py-8 relative">
-        <div className="flex my-6 gap-3">
-          <DetailTitle />
-
-          <h3>여행지코스 추천</h3>
-        </div>
-        <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
-          {displayedTravelCourse.map((item) => (
-            <TravelCard key={item.contentid} item={item} />
-          ))}
-        </div>
-        {displayedTravelCourse.length < filteredTravelCourse.length && (
-          <div ref={ref} className="py-4 text-center">
-            Loading more...
-          </div>
-        )}
-        <ScrollToTopButton />
-      </div>
-    </>
-  );
-};
-
-export default TravelCourse;
