--- src/app/api/main/Tour/AllFetch/leports/route.ts
@@ -1,9 +1,0 @@
-import { ApiInformation } from "@/types/Main";
-
-export const FetchLeports = async (): Promise<ApiInformation[]> => {
-  const res = await fetch("/api/main/Tour/leports");
-  if (!res.ok) {
-    throw new Error("error");
-  }
-  return res.json();
-};
