--- src/app/api/main/Tour/AllFetch/travel/route.ts
@@ -1,9 +1,0 @@
-import { ApiInformation } from "@/types/Main";
-
-export const FetchTravel = async (): Promise<ApiInformation[]> => {
-  const response = await fetch("/api/main/Tour/travel");
-  if (!response.ok) {
-    throw new Error("error");
-  }
-  return response.json();
-};
