+++ src/components/main/image/culturalFacilities/page.tsx
@@ -1,0 +1,7 @@
+import TourismList from "@/components/maindetail/TourismList";
+
+const culturalFacilities = () => {
+  return <TourismList contentTypeId={14} title="문화시설 추천" />;
+};
+
+export default culturalFacilities;
