+++ src/components/maindetail/TourismList.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Loading } from "@/components/common/Loading";
+import LoadingPage from "@/app/loading";
 import { DetailTitle } from "@/components/maindetail/DetailTitle";
 import { ScrollToTopButton } from "@/components/maindetail/ScrollToTopButton";
 import { SearchBar } from "@/components/maindetail/SearchBar";
@@ -10,50 +10,61 @@ import { useQuery } from "@tanstack/react-query";
 import { useState, useMemo, useEffect } from "react";
 import { useInView } from "react-intersection-observer";
 
-export const FetchLeports = async (): Promise<ApiInformation[]> => {
-  const res = await fetch("/api/main/Tour/leports");
-  if (!res.ok) {
-    throw new Error("error");
+// 관광 유형을 정의하는 인터페이스
+interface TourismListProps {
+  contentTypeId: number;
+  title: string;
+}
+
+// 관광 데이터를 가져오는 함수
+const fetchTourismData = async (contentTypeId: number): Promise<ApiInformation[]> => {
+  const response = await fetch(`/api/main/tourism?contentTypeId=${contentTypeId}`);
+  if (!response.ok) {
+    throw new Error(`Failed to fetch tourism data for contentTypeId ${contentTypeId}`);
   }
-  return res.json();
+  return response.json();
 };
 
-const Leports = () => {
+export const TourismList: React.FC<TourismListProps> = ({ contentTypeId, title }) => {
   const [displayCount, setDisplayCount] = useState(10);
   const [searchTerm, setSearchTerm] = useState("");
   const { ref, inView } = useInView();
 
+  // React Query를 사용하여 데이터 fetching
   const {
-    data: leports,
+    data: tourismData,
     isPending,
     error,
   } = useQuery<ApiInformation[], Error>({
-    queryKey: ["leports"],
-    queryFn: FetchLeports,
+    queryKey: ["tourism", contentTypeId],
+    queryFn: () => fetchTourismData(contentTypeId),
   });
 
-  const filteredLeports = useMemo(() => {
-    if (!leports) return [];
-    return leports
+  // 검색어에 따라 데이터 필터링 및 정렬
+  const filteredData = useMemo(() => {
+    if (!tourismData) return [];
+    return tourismData
       .filter((item) => item.title.toLowerCase().includes(searchTerm.toLowerCase()))
       .sort((a, b) => {
         if (a.firstimage && !b.firstimage) return -1;
         if (!a.firstimage && b.firstimage) return 1;
         return 0;
       });
-  }, [leports, searchTerm]);
+  }, [tourismData, searchTerm]);
 
-  const displayedLeports = useMemo(() => {
-    return filteredLeports.slice(0, displayCount);
-  }, [filteredLeports, displayCount]);
+  // 화면에 표시할 데이터
+  const displayedData = useMemo(() => {
+    return filteredData.slice(0, displayCount);
+  }, [filteredData, displayCount]);
 
+  // Infinite Scrolling을 위한 useEffect
   useEffect(() => {
     if (inView) {
       setDisplayCount((prevCount) => prevCount + 10);
     }
   }, [inView]);
 
-  if (isPending) return <Loading />;
+  if (isPending) return <LoadingPage />;
   if (error) return <div>Error: {error.message}</div>;
 
   return (
@@ -62,15 +73,14 @@ const Leports = () => {
       <div className="container mx-auto px-4 py-8 relative">
         <div className="flex my-6 gap-3">
           <DetailTitle />
-
-          <h3>액티비티 추천</h3>
+          <h3>{title}</h3>
         </div>
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
-          {displayedLeports.map((item) => (
+          {displayedData.map((item) => (
             <TravelCard key={item.contentid} item={item} />
           ))}
         </div>
-        {displayedLeports.length < filteredLeports.length && (
+        {displayedData.length < filteredData.length && (
           <div ref={ref} className="py-4 text-center">
             Loading more...
           </div>
@@ -80,5 +90,3 @@ const Leports = () => {
     </>
   );
 };
-
-export default Leports;
